+++ sdmx-dl-format-kryo/src/test/java/sdmxdl/kryo/KryoSerializationTest.java
@@ -5,13 +5,16 @@ import nbbrd.io.FileParser;
 import org.junit.Test;
 import sdmxdl.repo.SdmxRepository;
 import sdmxdl.samples.RepoSamples;
+import sdmxdl.web.SdmxWebMonitorReport;
+import sdmxdl.web.SdmxWebMonitorReports;
+import sdmxdl.web.SdmxWebStatus;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.time.Clock;
+import java.time.Duration;
 import java.time.Instant;
-import java.time.temporal.ChronoUnit;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -22,15 +25,25 @@ public class KryoSerializationTest {
     public void test() throws IOException {
         Instant now = Clock.systemDefaultZone().instant();
 
-        SdmxRepository data = RepoSamples.REPO
+        SdmxRepository repository = RepoSamples.REPO
                 .toBuilder()
-                .creationTime(now)
-                .expirationTime(now.plus(100, ChronoUnit.MILLIS))
+                .ttl(now, Duration.ofMillis(100))
                 .build();
 
-        assertThat(storeLoad(KryoSerialization.getRepositoryParser(), KryoSerialization.getRepositoryFormatter(), data))
-                .isEqualTo(data)
-                .isNotSameAs(data);
+        assertThat(storeLoad(KryoFileFormat.REPOSITORY, KryoFileFormat.REPOSITORY, repository))
+                .isEqualTo(repository)
+                .isNotSameAs(repository);
+
+        SdmxWebMonitorReports reports = SdmxWebMonitorReports
+                .builder()
+                .provider("abc")
+                .report(SdmxWebMonitorReport.builder().source("xyz").status(SdmxWebStatus.DOWN).uptimeRatio(0.5).averageResponseTime(1234L).build())
+                .ttl(now, Duration.ofMillis(100))
+                .build();
+
+        assertThat(storeLoad(KryoFileFormat.MONITOR, KryoFileFormat.MONITOR, reports))
+                .isEqualTo(reports)
+                .isNotSameAs(reports);
     }
 
     private static <T> T storeLoad(FileParser<T> parser, FileFormatter<T> formatter, T data) throws IOException {