+++ sdmx-dl-provider-util/src/main/java/sdmxdl/util/ext/MapCache.java
@@ -16,15 +16,18 @@
  */
 package sdmxdl.util.ext;
 
+import nbbrd.design.VisibleForTesting;
 import org.checkerframework.checker.nullness.qual.NonNull;
 import org.checkerframework.checker.nullness.qual.Nullable;
 import sdmxdl.ext.SdmxCache;
 import sdmxdl.repo.SdmxRepository;
+import sdmxdl.web.SdmxWebMonitorReports;
 
 import java.time.Clock;
 import java.util.Objects;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
+import java.util.function.Predicate;
 
 /**
  * @author Philippe Charles
@@ -35,39 +38,69 @@ public final class MapCache implements SdmxCache {
 
     @NonNull
     public static MapCache of() {
-        return of(new ConcurrentHashMap<>(), Clock.systemDefaultZone());
+        return of(new ConcurrentHashMap<>(), new ConcurrentHashMap<>(), Clock.systemDefaultZone());
     }
 
     @lombok.NonNull
-    private final ConcurrentMap<String, SdmxRepository> map;
+    private final ConcurrentMap<String, SdmxRepository> repositories;
+
+    @lombok.NonNull
+    private final ConcurrentMap<String, SdmxWebMonitorReports> webMonitors;
 
     @lombok.NonNull
     private final Clock clock;
 
     @Override
-    public SdmxRepository getRepository(String key) {
-        return get(map, clock, key);
+    public @Nullable SdmxRepository getRepository(@NonNull String key) {
+        Objects.requireNonNull(key);
+        return getRepository(repositories, clock, key);
     }
 
     @Override
-    public void putRepository(String key, SdmxRepository value) {
-        put(map, key, value);
+    public void putRepository(@NonNull String key, @NonNull SdmxRepository value) {
+        Objects.requireNonNull(key);
+        Objects.requireNonNull(value);
+        put(repositories, key, value);
+    }
+
+    @Override
+    public @Nullable SdmxWebMonitorReports getWebMonitorReports(@NonNull String key) {
+        Objects.requireNonNull(key);
+        return getWebMonitorReports(webMonitors, clock, key);
+    }
+
+    @Override
+    public void putWebMonitorReports(@NonNull String key, @NonNull SdmxWebMonitorReports value) {
+        Objects.requireNonNull(key);
+        Objects.requireNonNull(value);
+        put(webMonitors, key, value);
+    }
+
+    @VisibleForTesting
+    static SdmxRepository getRepository(@NonNull ConcurrentMap<String, SdmxRepository> map, @NonNull Clock clock, @NonNull String key) {
+        return get(reports -> !reports.isExpired(clock), map, key);
+    }
+
+    @VisibleForTesting
+    static SdmxWebMonitorReports getWebMonitorReports(@NonNull ConcurrentMap<String, SdmxWebMonitorReports> map, @NonNull Clock clock, @NonNull String key) {
+        return get(reports -> !reports.isExpired(clock), map, key);
     }
 
     @Nullable
-    static SdmxRepository get(@NonNull ConcurrentMap<String, SdmxRepository> map, @NonNull Clock clock, @NonNull String key) {
-        SdmxRepository value = map.get(key);
-        if (value == null) {
+    private static <T> T get(@NonNull Predicate<T> validator, @NonNull ConcurrentMap<String, T> map, @NonNull String key) {
+        T result = map.get(key);
+        if (result == null) {
             return null;
         }
-        if (value.isExpired(clock)) {
+        if (!validator.test(result)) {
             map.remove(key);
             return null;
         }
-        return value;
+        return result;
     }
 
-    static void put(@NonNull ConcurrentMap<String, SdmxRepository> map, @NonNull String key, @NonNull SdmxRepository value) {
+    @VisibleForTesting
+    static <T> void put(@NonNull ConcurrentMap<String, T> map, @NonNull String key, @NonNull T value) {
         Objects.requireNonNull(value);
         map.put(key, value);
     }