+++ sdmx-dl-provider-util/src/test/java/sdmxdl/util/ext/FileCacheTest.java
@@ -18,7 +18,8 @@ package sdmxdl.util.ext;
 
 import com.google.common.jimfs.Configuration;
 import com.google.common.jimfs.Jimfs;
-import org.checkerframework.checker.nullness.qual.NonNull;
+import nbbrd.io.FileFormatter;
+import nbbrd.io.FileParser;
 import org.junit.Test;
 import sdmxdl.repo.SdmxRepository;
 import sdmxdl.tck.ext.FakeClock;
@@ -50,13 +51,13 @@ public class FileCacheTest {
         try (FileSystem fs = Jimfs.newFileSystem(Configuration.unix())) {
             FakeClock clock = new FakeClock();
             clock.set(1000);
-            FakeSerializer serializer = new FakeSerializer();
+            FileFormat<SdmxRepository> serializer = newFakeFileFormat();
 
             FileCache cache = FileCache
                     .builder()
                     .root(fs.getPath("/").resolve("testfolder"))
                     .fileNameGenerator(UnaryOperator.identity())
-                    .serializer(serializer)
+                    .repositoryFormat(serializer)
                     .clock(clock)
                     .build();
 
@@ -72,7 +73,7 @@ public class FileCacheTest {
                     .ttl(clock.instant(), Duration.ofMillis(10))
                     .build();
             cache.putRepository("KEY1", r1);
-            assertThat(cache.getFile("KEY1"))
+            assertThat(cache.getFile("KEY1", FileCache.FileType.REPOSITORY, serializer))
                     .exists()
                     .hasContent("r1");
 
@@ -89,7 +90,7 @@ public class FileCacheTest {
             assertThat(cache.getRepository("KEY1"))
                     .as("Expired key should return null")
                     .isNull();
-            assertThat(cache.getFile("KEY1"))
+            assertThat(cache.getFile("KEY1", FileCache.FileType.REPOSITORY, serializer))
                     .as("Expired key should be deleted")
                     .doesNotExist();
 
@@ -104,7 +105,7 @@ public class FileCacheTest {
                     .build();
             cache.putRepository("KEY1", r1b);
             cache.putRepository("KEY1", r2);
-            assertThat(cache.getFile("KEY1"))
+            assertThat(cache.getFile("KEY1", FileCache.FileType.REPOSITORY, serializer))
                     .exists()
                     .hasContent("r2");
             assertThat(cache.getRepository("KEY1"))
@@ -113,25 +114,26 @@ public class FileCacheTest {
         }
     }
 
-    private static final class FakeSerializer implements Serializer<SdmxRepository> {
+    private static FileFormat<SdmxRepository> newFakeFileFormat() {
+        Map<String, SdmxRepository> content = new HashMap<>();
+        return new FileFormat<>(newFakeFileParser(content), newFakeFileFormatter(content), ".dat");
+    }
 
-        @lombok.Getter
-        private final Map<String, SdmxRepository> content = new HashMap<>();
-
-        @Override
-        public @NonNull SdmxRepository parseStream(@NonNull InputStream stream) throws IOException {
+    private static FileParser<SdmxRepository> newFakeFileParser(Map<String, SdmxRepository> content) {
+        return (stream) -> {
             try (Reader reader = new InputStreamReader(stream, StandardCharsets.UTF_8)) {
                 String name = new BufferedReader(reader).lines().collect(Collectors.joining(""));
                 return content.get(name);
             }
-        }
+        };
+    }
 
-        @Override
-        public void formatStream(@NonNull SdmxRepository entry, @NonNull OutputStream stream) throws IOException {
+    private static FileFormatter<SdmxRepository> newFakeFileFormatter(Map<String, SdmxRepository> content) {
+        return (entry, stream) -> {
             try (Writer writer = new OutputStreamWriter(stream, StandardCharsets.UTF_8)) {
                 writer.write(entry.getName());
                 content.put(entry.getName(), entry);
             }
-        }
+        };
     }
 }