+++ sdmx-dl-provider-util/src/test/java/sdmxdl/util/ext/MapCacheTest.java
@@ -16,7 +16,6 @@
  */
 package sdmxdl.util.ext;
 
-import _test.sdmxdl.util.CachingAssert;
 import org.junit.Test;
 import sdmxdl.repo.SdmxRepository;
 import sdmxdl.tck.ext.SdmxCacheAssert;
@@ -25,6 +24,7 @@ import java.time.Duration;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 
+import static _test.sdmxdl.util.CachingAssert.clock;
 import static org.assertj.core.api.Assertions.assertThat;
 
 /**
@@ -38,41 +38,41 @@ public class MapCacheTest {
     }
 
     @Test
-    public void testGet() {
+    public void testGetRepository() {
         ConcurrentMap<String, SdmxRepository> map = new ConcurrentHashMap<>();
 
-        assertThat(MapCache.get(map, CachingAssert.clock(1000), "KEY1"))
+        assertThat(MapCache.getRepository(map, clock(1000), "KEY1"))
                 .as("Empty map should return null")
                 .isNull();
 
         SdmxRepository r1000 = SdmxRepository
                 .builder()
                 .name("r1")
-                .ttl(CachingAssert.clock(1000).instant(), Duration.ofMillis(10))
+                .ttl(clock(1000).instant(), Duration.ofMillis(10))
                 .build();
         MapCache.put(map, "KEY1", r1000);
-        assertThat(MapCache.get(map, CachingAssert.clock(1009), "KEY1"))
+        assertThat(MapCache.getRepository(map, clock(1009), "KEY1"))
                 .as("Non-expired key should return value")
                 .isEqualTo(r1000);
 
-        assertThat(MapCache.get(map, CachingAssert.clock(1010), "KEY1"))
+        assertThat(MapCache.getRepository(map, clock(1010), "KEY1"))
                 .as("Expired key should return null")
                 .isNull();
         assertThat(map)
                 .as("Expired key should be evicted")
                 .doesNotContainKey("KEY1");
 
-        assertThat(MapCache.get(map, CachingAssert.clock(1009), "KEY2"))
+        assertThat(MapCache.getRepository(map, clock(1009), "KEY2"))
                 .as("Non-existing key should return null")
                 .isNull();
 
         SdmxRepository r1009 = SdmxRepository
                 .builder()
                 .name("r2")
-                .ttl(CachingAssert.clock(1009).instant(), Duration.ofMillis(10))
+                .ttl(clock(1009).instant(), Duration.ofMillis(10))
                 .build();
         MapCache.put(map, "KEY1", r1009);
-        assertThat(MapCache.get(map, CachingAssert.clock(1010), "KEY1"))
+        assertThat(MapCache.getRepository(map, clock(1010), "KEY1"))
                 .as("Updated key should return updated value")
                 .isEqualTo(r1009);
     }