+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Rename command check.properties as check.config [#136](https://github.com/nbbrd/sdmx-dl/issues/136)
 - Migrate web monitors to Upptime [#142](https://github.com/nbbrd/sdmx-dl/issues/142)
+- Refactor cache to handle web monitors
 
 ## [3.0.0-beta.3] - 2021-09-13